+++ History.md
@@ -2,6 +2,7 @@ Unpublished
 ===========
 
   * export Gpio constructor as exports.Gpio
+  * high precision notifications with Notifier
 
 0.0.8 / Nov 08 2015
 ===================